feat: migrate calculations from Double to BigDecimal for higher precision

+++ app/src/main/java/com/github/mykolab31/plugincalculator/data/model/CalculationResult.kt
@@ -1,7 +1,9 @@
 package com.github.mykolab31.plugincalculator.data.model
 
+import java.math.BigDecimal
+
 sealed class CalculationResult {
-    data class Number(val value: Double) : CalculationResult()
-    data class Matrix(val rows: List<List<Double>>) : CalculationResult()
+    data class Number(val value: BigDecimal) : CalculationResult()
+    data class Matrix(val rows: List<List<BigDecimal>>) : CalculationResult()
     data class Err(val message: String) : CalculationResult()
 }
